can upload

+++ src/app/component/content/content-detail/content-detail.component.ts
@@ -39,17 +39,71 @@ export class ContentDetailComponent implements OnInit {
   getImageUrl() {
     return "https://vkadmin.thantzin.pro/images/medium/"
   }
+  getOriginalImageUrl() {
+    return "https://vkadmin.thantzin.pro/images/original/"
+  }
   toggleEditMode() {
     this.editHide = !this.editHide
   }
+  private onSuccess() {
+    this.selectedFile.pending = false;
+    this.selectedFile.status = 'ok';
+  }
+
+  private onError() {
+    this.selectedFile.pending = false;
+    this.selectedFile.status = 'fail';
+    this.selectedFile.src = '';
+  }
+  processFile(event) {
+    const file: File = event.target.files[0]
+    const reader = new FileReader();
+    reader.addEventListener('load', (event: any) => {
+
+      this.selectedFile = new ImageSnippet(event.target.result, file);
+      this.contentService.uploadImage(this.contentId, file).subscribe(res => {
+        this.onSuccess();
+      }, (err) => {
+        this.onError();
+      })
+    });
+
+    reader.readAsDataURL(file);
+  }
   deleteContent() {
-    this.contentService.deleteContent(this.contentId).subscribe(res => {
-      if (res.status === EnumResponseStatus.Accepted) {
-        alert("successfully deleted")
+    const title = "Delete confirm"
+    const message = "Are you sure you want to delete the post? By clicking Yes you will lost the images."
+    this.dialog.open(ConfirmDialog, {
+      maxWidth: "400px",
+      data: {title: title, message: message}
+    }).afterClosed().subscribe(res => {
+      if (res) {
+        this.contentService.deleteContent(this.contentId).subscribe(res => {
+          if (res.status === EnumResponseStatus.Accepted) {
+            alert("successfully deleted")
+          }
+        })
       }
     })
   }
-  deleteContentImg() {
-
+  deleteContentImg(img: string) {
+    const title = "Delete confirm"
+    const message = "Are you sure you want to delete the image?"
+    this.dialog.open(ConfirmDialog, {
+      maxWidth: "400px",
+      data: {title: title, message: message}
+    }).afterClosed().subscribe(res => {
+      if (res) {
+        this.contentService.deleteImage(this.contentId, img).subscribe(result => {
+          if (result.status === EnumResponseStatus.Accepted) {
+            alert("The image is deleted")
+            this.populateImage(img)
+          }
+        })
+      }
+    })
+  }
+  populateImage(img: string) {
+    this.content.images = this.content.images.filter(x => x !== img)
   }
 }
